+++ tests/test_site.py
@@ -1,6 +1,5 @@
 from unittest import TestCase
 from staticsite.cmd.build import Build
-from staticsite import Site
 from . import utils as test_utils
 import os